move tests for IT to a separate file
