+++ src/test/java/ro/go/kpaxplanet/vat/validator/package-info.java
@@ -1,12 +1,12 @@
 /*******************************************************************************
  * Copyright 2015 Eugen Covaci
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,10 +14,10 @@
  * limitations under the License.
  ******************************************************************************/
 /**
- * 
+ *
  */
 /**
  * @author eugen
  *
  */
-package ro.go.kpaxplanet.vat.test;
+package ro.go.kpaxplanet.vat.validator;
